I need help pls
Hello everyone, I have a question and I would be very grateful if someone could help me.  I'm new to all this and I'm not doing very well. I have a modem from ISP in my house to which my Asus router is connected.  In the modem, Asus is set in the DMZ.  Modem model is ZTE F680 and its software is very old V4.0.10P3T4.  Please, if someone can help me how to find and download the latest firmware for that device.  Thank you very much